# Optimize for x86 AVX