--- a/dlk/python/dlk/templates/include/tensor_view.h
+++ b/dlk/python/dlk/templates/include/tensor_view.h
@@ -245,7 +245,7 @@ class TensorView<QuantizedPacked<T>, memory_layout> {
 
 #ifdef RUN_ON_FPGA
 using kernel_t = TensorView<QUANTIZED_PACKED_KERNEL, MemoryLayout::OhIhHWOlIl>;
-#elif defined USE_NEON
+#elif defined USE_NEON || defined USE_AVX
 using kernel_t = TensorView<QUANTIZED_PACKED_KERNEL, MemoryLayout::NHWC>;
 #else
 using kernel_t = TensorView<QUANTIZED_PACKED_KERNEL, MemoryLayout::HWNC>;