@ruimashita 2019-07-17
on `Intel(R) Xeon(R) CPU E5-2623 v4 @ 2.60GHz`,
TotalRunTime: 
* avx: `28.354ms@OMP_NUM_THREADS=4`
* no avx: `142.131ms`


Commands:
--

```
$ python dlk/python/dlk/scripts/generate_project.py -i dlk/examples/object_detection/widerface_v5/minimal_graph_with_shape.pb -o tmp/ -p object_detection -hq -ts
$ cd tmp/object_detection.prj
$ CXXFLAGS=-DFUNC_TIME_MEASUREMENT make -j8 lm_x86_avx
$ make clean
$ CXXFLAGS=-DFUNC_TIME_MEASUREMENT make -j8 lm_x86
$ tar -zxf ../../dlk/examples/object_detection/widerface_v5/network_input_output.tar.gz
```

run using avx
```
$ OMP_NUM_THREADS=4 ./lm_x86_avx.elf network_input_output/000_images_placeholder:0.npy network_input_output/337_output:0.npy
-------------------------------------------------------------
Comparison: Default network test  succeeded!!!
-------------------------------------------------------------
Add,13,3,  sum:0.016ms
ApplyThresholdsAndPack,183,56,17,14,19,10,11,8,10,16,16,9,13,10,  sum:0.392ms
BatchNorm,2819,34,  sum:2.853ms
Convolution,8801,417,  sum:9.218ms
ExtractImagePatches,347,95,14,11,4,1,4,14,  sum:0.49ms
Max,56,  sum:0.056ms
Memcpy,65,7,4,3,2,1,0,0,2,4,0,0,1,0,  sum:0.089ms
Output tensor convert,27,  sum:0.027ms
QTZ_linear_mid_tread_half,10629,  sum:10.629ms
Quantized Conv2D Tiling,1379,431,447,193,215,164,87,22,39,68,501,86,68,87,68,  sum:3.855ms
QuantizedConv2D,2057,519,480,216,243,180,102,35,58,92,524,99,86,102,103,  sum:4.896ms
QuantizedConv2D_ApplyScalingFactor,26,  sum:0.026ms
Tensor convert,442,17,8,4,1,1,0,1,3,2,2,0,0,0,0,  sum:0.481ms
TotalInitTime,15922,  sum:15.922ms
TotalRunTime,28354,  sum:28.354ms
copy,42,10,3,1,2,1,0,0,2,2,1,0,1,0,5,  sum:0.07ms
func_ConcatOnDepth,18,  sum:0.018ms
kn2row-1x1,8795,414,  sum:9.209ms
matrix_multiplication,8792,412,  sum:9.204ms
pack_input,7149,  sum:7.149ms
---------------------------------------------------
TotalInitTime 15921,  sum:15.921ms
TotalRunTime 28354,  sum:28.354ms
..Convolution 8801,416,  sum:9.217ms
....kn2row-1x1 8795,414,  sum:9.209ms
......matrix_multiplication 8791,411,  sum:9.202ms
..BatchNorm 2819,34,  sum:2.853ms
..QTZ_linear_mid_tread_half 10628,  sum:10.628ms
....pack_input 7149,  sum:7.149ms
..QuantizedConv2D 2057,519,480,216,243,180,102,35,58,92,524,99,86,102,103,  sum:4.896ms
....Tensor convert 442,16,8,4,1,1,0,1,2,2,1,0,0,0,0,  sum:0.478ms
....Quantized Conv2D Tiling 1379,431,447,193,215,164,86,21,39,68,501,86,68,87,68,  sum:3.853ms
....ApplyThresholdsAndPack 183,56,17,14,19,10,11,8,10,16,16,9,13,10,  sum:0.392ms
....copy 42,10,3,1,2,1,0,0,2,2,1,0,1,0,5,  sum:0.07ms
....Output tensor convert 26,  sum:0.026ms
..Memcpy 64,7,4,2,1,0,0,0,2,4,0,0,1,0,  sum:0.085ms
..ExtractImagePatches 347,95,14,11,4,1,4,14,  sum:0.49ms
..func_ConcatOnDepth 18,  sum:0.018ms
..QuantizedConv2D_ApplyScalingFactor 25,  sum:0.025ms
..Add 13,3,  sum:0.016ms
..Max 55,  sum:0.055ms

```


run no avx
```
$ ./lm_x86.elf network_input_output/000_images_placeholder:0.npy network_input_output/337_output:0.npy
-------------------------------------------------------------
Comparison: Default network test  succeeded!!!
-------------------------------------------------------------
Add,7,2,  sum:0.009ms
ApplyThresholds,4698,989,281,138,57,17,26,18,52,191,30,20,31,27,  sum:6.575ms
BatchNorm,3637,19,  sum:3.656ms
Convolution,10259,213,  sum:10.472ms
ExtractImagePatches,316,61,9,8,3,1,3,9,  sum:0.41ms
Max,31,  sum:0.031ms
Memcpy,69,15,4,2,1,0,1,0,1,3,0,0,1,1,  sum:0.098ms
QTZ_linear_mid_tread_half,13952,  sum:13.952ms
QuantizedConv2D,61463,14059,7819,6652,3514,2096,2543,228,325,716,10350,877,904,1000,879,  sum:113.425ms
QuantizedConv2D_ApplyScalingFactor,15,  sum:0.015ms
Tensor convert,59,93,33,16,8,4,1,4,9,16,7,2,1,2,2,  sum:0.257ms
TotalInitTime,19007,  sum:19.007ms
TotalRunTime,142132,  sum:142.132ms
func_ConcatOnDepth,12,  sum:0.012ms
kn2row-1x1,10252,211,  sum:10.463ms
matrix_multiplication,10247,209,  sum:10.456ms
matrix_shift_add1,971,293,139,118,109,54,104,104,53,92,52,92,  sum:2.181ms
matrix_shift_add2,6680,1099,272,97,38,10,21,18,10,17,11,17,  sum:8.29ms
pack bits,2645,510,125,56,29,7,14,7,29,112,14,7,13,7,  sum:3.575ms
pack_input,5917,  sum:5.917ms
quantized-kn2row,61395,13963,7784,6634,3503,2090,2540,222,314,698,10340,873,901,995,875,  sum:113.127ms
quantized_matrix_multiplication,38249,9421,6888,6182,3245,1990,2359,190,226,381,10159,772,725,873,716,  sum:82.376ms
---------------------------------------------------
TotalInitTime 19006,  sum:19.006ms
TotalRunTime 142131,  sum:142.131ms
..Convolution 10259,212,  sum:10.471ms
....kn2row-1x1 10252,211,  sum:10.463ms
......matrix_multiplication 10246,209,  sum:10.455ms
..BatchNorm 3636,18,  sum:3.654ms
..QTZ_linear_mid_tread_half 13951,  sum:13.951ms
....pack_input 5917,  sum:5.917ms
..QuantizedConv2D 61463,14059,7819,6651,3513,2096,2543,228,325,716,10349,876,904,1000,879,  sum:113.421ms
....Tensor convert 59,92,32,16,8,4,1,4,8,16,7,2,1,2,1,  sum:0.253ms
....quantized-kn2row 61395,13962,7784,6633,3503,2090,2540,222,314,698,10340,872,901,995,875,  sum:113.124ms
......quantized_matrix_multiplication 38248,9421,6888,6182,3245,1989,2358,190,226,381,10158,771,725,873,715,  sum:82.37ms
......matrix_shift_add1 970,293,139,118,109,54,103,104,53,92,52,92,  sum:2.179ms
......matrix_shift_add2 6679,1099,271,97,38,10,21,18,10,17,10,17,  sum:8.287ms
......ApplyThresholds 4698,989,280,138,57,17,26,18,52,191,29,20,31,27,  sum:6.573ms
......pack bits 2644,509,125,56,28,7,14,7,28,112,14,7,13,7,  sum:3.571ms
..Memcpy 68,15,3,1,1,0,0,0,1,3,0,0,1,1,  sum:0.094ms
..ExtractImagePatches 316,60,8,8,3,1,3,9,  sum:0.408ms
..func_ConcatOnDepth 12,  sum:0.012ms
..QuantizedConv2D_ApplyScalingFactor 14,  sum:0.014ms
..Add 7,2,  sum:0.009ms
..Max 30,  sum:0.03ms

```